Allow PrivateNetworkOnlyFlag option
This commit adds the option in softlayer packer
plugin to set up an instance with private only
network.
Allows to specify private_network_only_flag parameter,
when set to true, the instance spun up will only
have a private ip. And it will try to ssh and run the
scripts using this private ip.